+++ modules/gltf/src/lib/extensions/EXT_texture_webp.ts
@@ -5,7 +5,7 @@
 import type {GLTF, GLTF_EXT_texture_webp} from '../types/gltf-types';
 import type {GLTFLoaderOptions} from '../../gltf-loader';
 
-import {_isImageFormatSupported} from '@loaders.gl/images';
+import {isImageFormatSupported} from '@loaders.gl/images';
 import GLTFScenegraph from '../api/gltf-scenegraph';
 
 const EXT_TEXTURE_WEBP = 'EXT_texture_webp';
@@ -20,7 +20,7 @@ export const name = EXT_TEXTURE_WEBP;
 export function preprocess(gltfData: {json: GLTF}, options: GLTFLoaderOptions): void {
   const scenegraph = new GLTFScenegraph(gltfData);
 
-  if (!_isImageFormatSupported('image/webp')) {
+  if (!isImageFormatSupported('image/webp')) {
     if (scenegraph.getRequiredExtensions().includes(EXT_TEXTURE_WEBP)) {
       throw new Error(`gltf: Required extension ${EXT_TEXTURE_WEBP} not supported by browser`);
     }